[cleanup] abstract detail rows
